Remove spurious tests, refactor
